+++ commands/audit.go
@@ -28,7 +28,13 @@ import (
 	"fmt"
 	"io"
 	"net/http"
+	"sort"
+	"strings"
+	"time"
 
+	"github.com/briandowns/spinner"
+	"github.com/fatih/color"
+	"github.com/jedib0t/go-pretty/v6/table"
 	"github.com/mkloubert/go-package-manager/types"
 	"github.com/mkloubert/go-package-manager/utils"
 	"github.com/spf13/cobra"
@@ -40,6 +46,8 @@ func Init_Audit_Command(parentCmd *cobra.Command, app *types.AppContext) {
 		Short: "Audit modules",
 		Long:  `Audits modules of the current project using API of osv.dev`,
 		Run: func(cmd *cobra.Command, args []string) {
+			tHeadColor := color.New(color.FgWhite, color.Bold).SprintFunc()
+
 			modules, err := app.GetGoModules()
 			if err != nil {
 				utils.CloseWithError(err)
@@ -59,18 +67,27 @@ func Init_Audit_Command(parentCmd *cobra.Command, app *types.AppContext) {
 						return
 					}
 
-					logAuditError := func(step string, err error) {
-						app.L.Printf("ERROR (%v): Could not audit module #%v (%v %v): '%v'%v",
-							step,
-							i,
-							*modulePath, *moduleVersion,
-							err,
+					coloredModuleName := color.New(color.FgWhite, color.Bold).Sprint(*modulePath)
+					coloredModuleVersion := color.New(color.FgWhite, color.Bold).Sprint(*moduleVersion)
+
+					s := spinner.New(spinner.CharSets[24], 100*time.Millisecond)
+					s.Start()
+					s.Suffix = fmt.Sprintf(
+						" %v (%v)",
+						coloredModuleName, coloredModuleVersion,
+					)
+					s.Color("white")
+
+					stopByError := func(err error) {
+						s.Stop()
+
+						fmt.Printf(
+							"❌ %v (%v): %v%v",
+							coloredModuleName, coloredModuleVersion,
+							color.New(color.FgYellow, color.BgRed, color.Bold).Sprint(err),
 							fmt.Sprintln(),
 						)
 					}
-
-					// now we have all information
-					fmt.Printf("%v\t%v%v", *modulePath, *moduleVersion, fmt.Sprintln())
 
 					// prepare request to osv.dev API
 					url := "https://api.osv.dev/v1/query"
@@ -85,23 +102,14 @@ func Init_Audit_Command(parentCmd *cobra.Command, app *types.AppContext) {
 					// serialize body
 					jsonData, err := json.Marshal(&body)
 					if err != nil {
-						logAuditError("create request body", err)
+						stopByError(fmt.Errorf("could not serialize request body: %v", err))
 						return
 					}
-
-					app.Debug(
-						fmt.Sprintf(
-							"Doing POST request to '%v' for module #%v (%v %v) ...",
-							url,
-							i,
-							*modulePath, *moduleVersion,
-						),
-					)
 
 					// start the request
 					req, err := http.NewRequest("POST", url, bytes.NewBuffer([]byte(jsonData)))
 					if err != nil {
-						logAuditError("create POST request", err)
+						stopByError(fmt.Errorf("could not prepare POST request to '%v': %v", url, err))
 						return
 					}
 
@@ -111,15 +119,20 @@ func Init_Audit_Command(parentCmd *cobra.Command, app *types.AppContext) {
 					client := &http.Client{}
 					resp, err := client.Do(req)
 					if err != nil {
-						logAuditError("do POST request", err)
+						stopByError(fmt.Errorf("could not do POST request to '%v': %v", url, err))
 						return
 					}
 					defer resp.Body.Close()
 
+					if resp.StatusCode != 200 {
+						stopByError(fmt.Errorf("unexpected response from '%v': %v", url, resp.StatusCode))
+						return
+					}
+
 					// load the response
 					osvResponseData, err := io.ReadAll(resp.Body)
 					if err != nil {
-						logAuditError("read response data", err)
+						stopByError(fmt.Errorf("could not do load response from '%v': %v", url, err))
 						return
 					}
 
@@ -127,65 +140,125 @@ func Init_Audit_Command(parentCmd *cobra.Command, app *types.AppContext) {
 					var osvResponse types.OsvDevResponse
 					err = json.Unmarshal(osvResponseData, &osvResponse)
 					if err != nil {
-						logAuditError("unmarshal response data", err)
+						stopByError(fmt.Errorf("could not parse response from '%v': %v", url, err))
 						return
+					}
+
+					s.Stop()
+
+					printNoIssueInfo := func() {
+						fmt.Printf(
+							"✅ %v (%v)%v",
+							coloredModuleName, coloredModuleVersion,
+							fmt.Sprintln(),
+						)
 					}
 
 					if osvResponse.Vulnerabilities == nil {
-						// we have no data here
-
-						app.Debug(
-							fmt.Sprintf(
-								"Found no vulnerabilities for module #%v (%v %v) ...",
-								i,
-								*modulePath, *moduleVersion,
-							),
-						)
+						printNoIssueInfo()
 						return
 					}
 
-					vulnerabilities := *osvResponse.Vulnerabilities
+					// create copy of array in osvResponse.Vulnerabilities ...
+					vulnerabilities := []types.OsvDevResponseVulnerabilityItem{}
+					vulnerabilities = append(vulnerabilities, *osvResponse.Vulnerabilities...)
 					vulnerabilitiesCount := len(vulnerabilities)
 
-					app.Debug(
-						fmt.Sprintf(
-							"Found %v vulnerabilities for module #%v (%v %v) ...",
-							vulnerabilitiesCount,
-							i,
-							*modulePath, *moduleVersion,
-						),
+					if vulnerabilitiesCount == 0 {
+						printNoIssueInfo()
+						return
+					}
+
+					// sort by severity (desc)
+					sort.Slice(vulnerabilities, func(x int, y int) bool {
+						vulnX := vulnerabilities[x]
+						vulnY := vulnerabilities[y]
+
+						_, compX := vulnX.GetSeverityDisplayValues()
+						_, compY := vulnY.GetSeverityDisplayValues()
+						if compX != compY {
+							return compX > compY
+						}
+
+						return false
+					})
+
+					fmt.Printf(
+						"⚠️ %v (%v):%v",
+						coloredModuleName, coloredModuleVersion,
+						fmt.Sprintln(),
 					)
 
-					for _, v := range vulnerabilities {
-						summaryJsonData, err := json.Marshal(v.Summary)
-						if err != nil {
-							summaryJsonData = []byte{}
+					var tBuffer bytes.Buffer
+
+					// output in buffer first
+					t := table.NewWriter()
+					t.SetOutputMirror(&tBuffer)
+
+					// header
+					t.AppendHeader(table.Row{tHeadColor("#"), tHeadColor("Severity"), tHeadColor("ID"), tHeadColor("Summary")})
+					for vi, v := range vulnerabilities {
+						if vi > 0 {
+							// add separator at top
+							t.AppendSeparator()
 						}
 
-						// output [<Severity>] <Id> <Summary>
-						// with 1 leading tab
-						fmt.Printf(
-							"\t[%v]\t%v\t%v%v",
-							v.Severity,
-							v.Id,
-							string(summaryJsonData),
-							fmt.Sprintln(),
-						)
+						severity, _ := v.GetSeverityDisplayValues()
 
-						if v.References == nil {
-							continue
+						// output basic issue info
+						t.AppendRow(table.Row{vi + 1, severity, v.Id, v.Summary})
+
+						if v.References != nil {
+							// add references
+
+							references := []types.OsvDevResponseVulnerabilityItemReference{}
+							references = append(references, *v.References...)
+
+							// sort references by type, then by URL
+							sort.Slice(references, func(x int, y int) bool {
+								refX := references[x]
+								refY := references[y]
+
+								typeX := strings.TrimSpace(strings.ToLower(refX.Type))
+								typeY := strings.TrimSpace(strings.ToLower(refY.Type))
+								if typeX != typeY {
+									return typeX < typeY
+								}
+
+								urlX := strings.TrimSpace(strings.ToLower(refX.Url))
+								urlY := strings.TrimSpace(strings.ToLower(refY.Url))
+
+								return urlX < urlY
+							})
+
+							if len(references) > 0 {
+								// build reference list
+
+								t.AppendSeparator()
+
+								for ri, r := range references {
+									refCol := ""
+									if ri == 0 {
+										refCol = tHeadColor("References:")
+									}
+
+									t.AppendRow(table.Row{"", refCol, r.Type, r.Url})
+								}
+
+								t.AppendSeparator()
+							}
 						}
+					}
 
-						// output references
-						for _, r := range *v.References {
-							// output [<Type>] <Url>
-							// with 2 leading tabs
-							fmt.Printf(
-								"\t\t[%v]\t%v%v",
-								r.Type,
-								r.Url,
-								fmt.Sprintln(),
-							)
+					// render final table
+					t.Render()
+
+					// output final table with prefix
+					prefix := "  "
+					output := tBuffer.String()
+					for _, line := range strings.Split(output, fmt.Sprintln()) {
+						if len(line) > 0 {
+							fmt.Printf("%v%s%v", prefix, line, fmt.Sprintln())
 						}
 					}
 				}()